+++ gems/aws-sdk-mediatailor/lib/aws-sdk-mediatailor.rb
@@ -42,6 +42,6 @@ require_relative 'aws-sdk-mediatailor/customizations'
 # @service
 module Aws::MediaTailor
 
-  GEM_VERSION = '1.22.0'
+  GEM_VERSION = '1.23.0'
 
 end