+++ gems/aws-sdk-mediatailor/lib/aws-sdk-mediatailor/types.rb
@@ -189,6 +189,11 @@ module Aws::MediaTailor
     #   Amazon CloudFront, for content and ad segment management.
     #   @return [Types::CdnConfiguration]
     #
+    # @!attribute [rw] personalization_threshold_seconds
+    #   The maximum duration of underfilled ad time (in seconds) allowed in
+    #   an ad break.
+    #   @return [Integer]
+    #
     # @!attribute [rw] dash_configuration
     #   The configuration for DASH content.
     #   @return [Types::DashConfiguration]
@@ -250,6 +255,7 @@ module Aws::MediaTailor
     class GetPlaybackConfigurationResponse < Struct.new(
       :ad_decision_server_url,
       :cdn_configuration,
+      :personalization_threshold_seconds,
       :dash_configuration,
       :hls_configuration,
       :live_pre_roll_configuration,
@@ -355,6 +361,9 @@ module Aws::MediaTailor
     #   Amazon CloudFront, for content and ad segment management.
     #   @return [Types::CdnConfiguration]
     #
+    # @!attribute [rw] personalization_threshold_seconds
+    #   @return [Integer]
+    #
     # @!attribute [rw] dash_configuration
     #   The configuration for DASH content.
     #   @return [Types::DashConfiguration]
@@ -392,6 +401,7 @@ module Aws::MediaTailor
     class PlaybackConfiguration < Struct.new(
       :ad_decision_server_url,
       :cdn_configuration,
+      :personalization_threshold_seconds,
       :dash_configuration,
       :hls_configuration,
       :name,
@@ -448,6 +458,7 @@ module Aws::MediaTailor
     #           ad_segment_url_prefix: "__string",
     #           content_segment_url_prefix: "__string",
     #         },
+    #         personalization_threshold_seconds: 1,
     #         dash_configuration: {
     #           mpd_location: "__string",
     #           origin_manifest_type: "SINGLE_PERIOD", # accepts SINGLE_PERIOD, MULTI_PERIOD
@@ -478,6 +489,11 @@ module Aws::MediaTailor
     #   The configuration for using a content delivery network (CDN), like
     #   Amazon CloudFront, for content and ad segment management.
     #   @return [Types::CdnConfiguration]
+    #
+    # @!attribute [rw] personalization_threshold_seconds
+    #   The maximum duration of underfilled ad time (in seconds) allowed in
+    #   an ad break.
+    #   @return [Integer]
     #
     # @!attribute [rw] dash_configuration
     #   The configuration for DASH content.
@@ -522,6 +538,7 @@ module Aws::MediaTailor
     class PutPlaybackConfigurationRequest < Struct.new(
       :ad_decision_server_url,
       :cdn_configuration,
+      :personalization_threshold_seconds,
       :dash_configuration,
       :live_pre_roll_configuration,
       :name,